requirements.txt
asgiref==3.6.0
click==8.1.3
cmake==3.26.3
colorama==0.4.6
Django==4.2.1
djangorestframework==3.14.0
dlib @ https://github.com/jloh02/dlib/releases/download/v19.22/dlib-19.22.99-cp310-cp310-win_amd64.whl#sha256=f2181b6724669fb0147c9ebc326d72b94de41bced1812eff5df4b133b5d0b575
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.24.3
opencv-python==4.7.0.72
Pillow==9.5.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
whitenoise==6.4.0
